+++ model/build_graph.py
@@ -1,7 +1,6 @@
-import networkx as nx
 import scipy.sparse as sp
 from math import log
-from dataset import TextDataset
+from loader.dataset import TextDataset
 from collections import defaultdict
 import pandas as pd
 from utils import get_corpus_path
